Middlewares integration. Implementation of updateUser method. Add logout endpoint. Add forbidden error.

+++ server/src/services/UserService.js
@@ -1,21 +1,24 @@
 import User from "../models/User.js";
 import {comparePasswords, generateAccessToken, generateRefreshToken, hashPassword} from "../utils/jwt.js";
 import Session from "../models/Session.js";
+import {UserDto} from "../dtos/UserDto.js";
+import {ApiError} from "../exceptions/ApiError.js";
 
 
 class UserService {
 
     async login({email, password}) {
         if (!email || !password) {
-            throw new Error("Email and password are required");
+            throw ApiError.BadRequest("Необхідно ввести електронну пошту та пароль");
+
         }
 
-        const user = await User.findOne({email: email.toLowerCase()});
+        const user =  await User.findOne({email: email.toLowerCase()});
         if (!user) {
-            throw new Error("User not found");
+            throw ApiError.BadRequest("Користувача не знайдено");
         }
         if (!comparePasswords(password, user.password)) {
-            throw new Error("Password is incorrect");
+            throw ApiError.BadRequest("Невірний пароль");
         }
 
         return this.createSession(user);
@@ -23,21 +26,34 @@ class UserService {
 
     async register({email, password, name, surname}) {
         if (!email || !password || !name || !surname) {
-            throw new Error("Email, password, name and surname are required");
+            throw ApiError.BadRequest("Необхідно ввести електронну пошту, пароль, ім'я та прізвище");
         }
 
-        const existingUser = await User.findOne({email});
+        const existingUser =  await User.findOne({email: email.toLowerCase()})
         if (existingUser) {
-            throw new Error("User already exists");
+            throw ApiError.BadRequest("Користувач вже існує");
         }
+
+        //TODO generate activation link
+        //TODO send mail
+
         const newUser = await this.createUser({email, password, name, surname});
         return this.createSession(newUser);
     }
 
+    async logout(refreshToken) {
+        if (!refreshToken) {
+            throw ApiError.UnauthorizedError();
+        }
+        return Session.deleteOne({refresh_token: refreshToken});
+    }
+
+
     async createSession(user) {
+        const userDto = new UserDto(user);
         const tokens = {
-            access_token: generateAccessToken(user),
-            refresh_token: generateRefreshToken(user)
+            access_token: generateAccessToken({...userDto}),
+            refresh_token: generateRefreshToken({...userDto})
         }
 
         const existingSessions = await Session.find({user_id: user._id});
@@ -50,16 +66,15 @@ class UserService {
             refresh_token: tokens.refresh_token,
             expires_in: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30)
         });
-        return tokens;
+        return {
+            ...tokens,
+            user: userDto
+        };
     }
 
 
     async createUser(user) {
         const {email, password, name, surname} = user;
-        const existingUser = await User.findOne({email});
-        if (existingUser) {
-            throw new Error("User already exists");
-        }
 
         return User.create({
             email: email.toLowerCase(),
@@ -70,29 +85,61 @@ class UserService {
     }
 
     async getAllUsers() {
-        return User.find({});
+        const users = await User.find({});
+        return users.map(user => new UserDto(user));
     }
 
     async getUserById(id) {
-        return User.findOne({_id: id});
+        const user = await User.findOne({_id: id});
+        return new UserDto(user);
     }
 
     async getUserByEmail(email) {
-        return User.findOne({email: email.toLowerCase()});
+        const user = await User.findOne({email: email.toLowerCase()});
+        return new UserDto(user);
     }
 
 
-    async updateUser(id, user) {
-        const {email, password, name, surname} = user;
-        return User.findOneAndUpdate({_id: id}, {
-            email: email.toLowerCase(),
-            password: hashPassword(password),
-            name,
-            surname
-        }, {new: true});
+    async updateUser(id, targetId, user) {
+        const targetUser =  await User.findOne({_id: targetId});
+        if (!targetUser) {
+            throw ApiError.BadRequest("Користувача не знайдено");
+        }
+        if (id !== targetId) {
+            throw ApiError.ForbiddenError();
+        }
+
+        const updateData = {};
+        if(user.email){
+            const existingUser =  await User.findOne({email: user.email.toLowerCase()});
+            if (existingUser) {
+                throw ApiError.BadRequest("Адрес електронної пошти вже занято");
+            }
+            updateData.email = user.email.toLowerCase();
+            updateData.is_activated = false;
+            //TODO generate activation link
+            //TODO send mail
+        }
+        if (user.password) {
+            updateData.password = hashPassword(user.password);
+        }
+        if(user.name) {
+            updateData.name = user.name;
+        }
+        if(user.surname) {
+            updateData.surname = user.surname;
+        }
+        if(user.avatar) {
+            updateData.avatar = user.avatar;
+        }
+
+        return User.findOneAndUpdate({_id: targetId}, updateData, {new: true});
     }
 
-    async deleteUser(id) {
+    async deleteUser(id, targetId) {
+        if (id !== targetId) {
+            throw ApiError.ForbiddenError();
+        }
         return User.findOneAndDelete({_id: id});
     }
 }
